refactor(puppeteer): relaunch feature
